+++ site/src/pages/CreateUserPage/CreateUserForm.tsx
@@ -3,7 +3,8 @@ import { hasApiFieldErrors, isApiError } from "api/errors";
 import type * as TypesGen from "api/typesGenerated";
 import { ErrorAlert } from "components/Alert/ErrorAlert";
 import { Button } from "components/Button/Button";
-import { FormField, FormFooter } from "components/Form/Form";
+import { FormField } from "components/FormField/FormField";
+import { FormFooter } from "components/Form/Form";
 import { FullPageForm } from "components/FullPageForm/FullPageForm";
 import { Label } from "components/Label/Label";
 import { Link } from "components/Link/Link";
